# backend/docs/swagger.yaml
basePath: /api
definitions:
  db.CalculationFormularModel:
    properties:
      calculation:
        $ref: '#/definitions/db.CalculationModel'
      calculationId:
        type: string
      createdAt:
        type: string
      formular:
        $ref: '#/definitions/db.FormularModel'
      formularId:
        type: string
      id:
        type: string
      next:
        $ref: '#/definitions/db.CalculationFormularModel'
      nextId:
        type: string
      previous:
        $ref: '#/definitions/db.CalculationFormularModel'
      updatedAt:
        type: string
    type: object
  db.CalculationModel:
    properties:
      createdAt:
        type: string
      formulars:
        items:
          $ref: '#/definitions/db.CalculationFormularModel'
        type: array
      id:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
  db.FormularModel:
    properties:
      calculationFormulars:
        items:
          $ref: '#/definitions/db.CalculationFormularModel'
        type: array
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      nodes:
        items:
          $ref: '#/definitions/db.FormularNodeModel'
        type: array
      updatedAt:
        type: string
    type: object
  db.FormularNodeModel:
    properties:
      createdAt:
        type: string
      formular:
        $ref: '#/definitions/db.FormularModel'
      formularId:
        type: string
      id:
        type: string
      next:
        $ref: '#/definitions/db.FormularNodeModel'
      nextId:
        type: string
      node:
        $ref: '#/definitions/db.NodeModel'
      nodeId:
        type: string
      previous:
        $ref: '#/definitions/db.FormularNodeModel'
      updatedAt:
        type: string
    type: object
  db.NodeModel:
    properties:
      createdAt:
        type: string
      formularNodes:
        items:
          $ref: '#/definitions/db.FormularNodeModel'
        type: array
      id:
        type: string
      name:
        type: string
      nodeData:
        type: string
      updatedAt:
        type: string
    type: object
  handlers.AddFormularInput:
    properties:
      formularId:
        description: The ID of the formular to add
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      nextId:
        description: Optional ID of the next formular in sequence
        example: 123e4567-e89b-12d3-a456-426614174001
        type: string
    type: object
...
  handlers.CreateCalculationInput:
    properties:
      name:
        description: The name of the calculation
        example: My Calculation
        type: string
    type: object
...
  handlers.ReorderFormularsInput:
    properties:
      formularOrder:
        description: The ordered list of formular IDs
        example:
        - '[''123e4567-e89b-12d3-a456-426614174000'''
        - ' ''123e4567-e89b-12d3-a456-426614174001'']'
        items:
          type: string
        type: array
    type: object
...
  handlers.UpdateCalculationInput:
    properties:
      name:
        description: The new name of the calculation
        example: Updated Calculation
        type: string
    type: object
...
host: localhost:8081
info:
  contact: {}
  description: API for managing calculations, formulars, and nodes
  title: Calculation API
  version: "1.0"
paths:
  /calculations:
    get:
      consumes:
      - application/json
      description: Get all calculations
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.CalculationModel'
            type: array
      summary: List calculations
      tags:
      - calculations
    post:
      consumes:
      - application/json
      description: Create a new calculation
      parameters:
      - description: Calculation to create
        in: body
        name: calculation
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateCalculationInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/db.CalculationModel'
      summary: Create a calculation
      tags:
      - calculations
  /calculations/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a calculation by ID
      parameters:
      - description: Calculation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Calculation not found
          schema:
            type: string
      summary: Delete a calculation
      tags:
      - calculations
    get:
      consumes:
      - application/json
      description: Get calculation by ID
      parameters:
      - description: Calculation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.CalculationModel'
        "404":
          description: Calculation not found
          schema:
            type: string
      summary: Get a calculation
      tags:
      - calculations
    put:
      consumes:
      - application/json
      description: Update a calculation by ID
      parameters:
      - description: Calculation ID
        in: path
        name: id
        required: true
        type: string
      - description: Calculation updates
        in: body
        name: calculation
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateCalculationInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.CalculationModel'
        "404":
          description: Calculation not found
          schema:
            type: string
      summary: Update a calculation
      tags:
      - calculations
  /calculations/{id}/formulars:
    get:
      consumes:
      - application/json
      description: Get all formulars in a calculation's sequence
      parameters:
      - description: Calculation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.CalculationFormularModel'
            type: array
      summary: List formulars in a calculation
      tags:
      - calculations
    post:
      consumes:
      - application/json
      description: Add a formular to a calculation's sequence
      parameters:
      - description: Calculation ID
        in: path
        name: id
        required: true
        type: string
      - description: Formular to add
        in: body
        name: formular
        required: true
        schema:
          $ref: '#/definitions/handlers.AddFormularInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/db.CalculationFormularModel'
      summary: Add a formular to a calculation
      tags:
      - calculations
  /calculations/{id}/formulars/{formularId}:
    delete:
      consumes:
      - application/json
      description: Remove a formular from a calculation's sequence
      parameters:
      - description: Calculation ID
        in: path
        name: id
        required: true
        type: string
      - description: Formular ID
        in: path
        name: formularId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: CalculationFormular not found
          schema:
            type: string
      summary: Remove a formular from a calculation
      tags:
      - calculations
  /calculations/{id}/formulars/reorder:
    put:
      consumes:
      - application/json
      description: Update the sequence of formulars in a calculation
      parameters:
      - description: Calculation ID
        in: path
        name: id
        required: true
        type: string
      - description: New formular order
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/handlers.ReorderFormularsInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Reorder formulars in a calculation
      tags:
      - calculations
...
swagger: "2.0"